remove claimtxgasused and claimtxgasprice db columns, refactor metricsservice design pattern
