+++ postman/src/application/postman/persistence/mappers/__tests__/messageMappers.test.ts
@@ -22,8 +22,6 @@ describe("Message Mappers", () => {
         claimNumberOfRetry: 0,
         claimTxCreationDate: undefined,
         claimTxGasLimit: undefined,
-        claimTxGasUsed: undefined,
-        claimTxGasPrice: undefined,
         claimTxHash: undefined,
         claimTxMaxFeePerGas: undefined,
         claimTxMaxPriorityFeePerGas: undefined,
@@ -59,8 +57,6 @@ describe("Message Mappers", () => {
           claimNumberOfRetry: 0,
           claimTxCreationDate: undefined,
           claimTxGasLimit: undefined,
-          claimTxGasUsed: undefined,
-          claimTxGasPrice: undefined,
           claimTxHash: undefined,
           claimTxMaxFeePerGas: undefined,
           claimTxMaxPriorityFeePerGas: undefined,
