+++ postman/src/core/entities/Message.ts
@@ -16,8 +16,6 @@ export type MessageProps = {
   status: MessageStatus;
   claimTxCreationDate?: Date;
   claimTxGasLimit?: number;
-  claimTxGasUsed?: number;
-  claimTxGasPrice?: bigint;
   claimTxMaxFeePerGas?: bigint;
   claimTxMaxPriorityFeePerGas?: bigint;
   claimTxNonce?: number;
@@ -69,8 +67,6 @@ export class Message {
   public status: MessageStatus;
   public claimTxCreationDate?: Date;
   public claimTxGasLimit?: number;
-  public claimTxGasUsed?: number;
-  public claimTxGasPrice?: bigint;
   public claimTxMaxFeePerGas?: bigint;
   public claimTxMaxPriorityFeePerGas?: bigint;
   public claimTxNonce?: number;
@@ -98,8 +94,6 @@ export class Message {
     this.status = props.status;
     this.claimTxCreationDate = props.claimTxCreationDate;
     this.claimTxGasLimit = props.claimTxGasLimit;
-    this.claimTxGasUsed = props.claimTxGasUsed;
-    this.claimTxGasPrice = props.claimTxGasPrice;
     this.claimTxMaxFeePerGas = props.claimTxMaxFeePerGas;
     this.claimTxMaxPriorityFeePerGas = props.claimTxMaxPriorityFeePerGas;
     this.claimTxNonce = props.claimTxNonce;
@@ -121,8 +115,6 @@ export class Message {
     if (newMessage.status) this.status = newMessage.status;
     if (newMessage.claimTxCreationDate) this.claimTxCreationDate = newMessage.claimTxCreationDate;
     if (newMessage.claimTxGasLimit) this.claimTxGasLimit = newMessage.claimTxGasLimit;
-    if (newMessage.claimTxGasUsed) this.claimTxGasUsed = newMessage.claimTxGasUsed;
-    if (newMessage.claimTxGasPrice) this.claimTxGasPrice = newMessage.claimTxGasPrice;
     if (newMessage.claimTxMaxFeePerGas) this.claimTxMaxFeePerGas = newMessage.claimTxMaxFeePerGas;
     if (newMessage.claimTxMaxPriorityFeePerGas)
       this.claimTxMaxPriorityFeePerGas = newMessage.claimTxMaxPriorityFeePerGas;
@@ -147,9 +139,9 @@ export class Message {
       this.direction
     }, status=${this.status}, claimTxCreationDate=${this.claimTxCreationDate?.toISOString()}, claimTxGasLimit=${
       this.claimTxGasLimit
-    }, claimTxGasUsed=${this.claimTxGasUsed}, claimTxMaxFeePerGas=${this.claimTxMaxFeePerGas}, claimTxMaxPriorityFeePerGas=${
+    }, claimTxMaxFeePerGas=${this.claimTxMaxFeePerGas}, claimTxMaxPriorityFeePerGas=${
       this.claimTxMaxPriorityFeePerGas
-    }, claimTxGasPrice=${this.claimTxGasPrice}, claimTxNonce=${this.claimTxNonce}, claimTransactionHash=${this.claimTxHash}, claimNumberOfRetry=${
+    }, claimTxNonce=${this.claimTxNonce}, claimTransactionHash=${this.claimTxHash}, claimNumberOfRetry=${
       this.claimNumberOfRetry
     }, claimLastRetriedAt=${this.claimLastRetriedAt?.toISOString()}, claimGasEstimationThreshold=${
       this.claimGasEstimationThreshold
