+++ postman/src/core/metrics/MessageMetricsAttributes.ts
@@ -1,0 +1,8 @@
+import { MessageStatus } from "../enums";
+import { Direction } from "@consensys/linea-sdk";
+
+export type MessagesMetricsAttributes = {
+  status: MessageStatus;
+  direction: Direction;
+  isForSponsorship: boolean;
+};
