+++ postman/src/core/metrics/index.ts
@@ -1,0 +1,3 @@
+export * from "./IMessageMetricsUpdater";
+export * from "./IMetricsService";
+export * from "./MessageMetricsAttributes";
